refactor(clients): move bandcamp + streaming HTTP clients out of scripts/ into clients/ package
Production code under release/ and streaming/ was importing BandcampClient, SpotifyClient, DeezerClient, AppleMusicClient, BaseStreamingClient, and the streaming text-matching helpers from scripts/, treating CLI / batch-tooling modules as the source of truth for FastAPI dependencies.

The git mv targets: scripts/bandcamp_client.py becomes clients/bandcamp.py; scripts/streaming_availability/{spotify_client,deezer_client,apple_music_client,http_client,matching}.py become clients/streaming/{spotify,deezer,apple_music,base,matching}.py respectively. History preserved via git mv.

Bandcamp shares the same BaseStreamingClient lineage (rate-limited httpx.AsyncClient + lifecycle singleton), so it lives in clients/ next to the streaming subpackage. matching.py is a streaming-specific text-normalization / scoring helper coupled tightly to the client request shape and lives under clients/streaming/ for the same reason.

Rewrite every import site, including production files under release/ and streaming/, plus the scripts/ siblings (streaming_availability/{dedup,discogs_enricher,__main__}, track_streaming/*, bandcamp_pipeline, discogs_rematch, recheck_streaming_after_mojibake, revalidate_misses, search_unmatched_compilations) that still depend on the moved modules. Tests and monkeypatch target strings updated in lockstep.

Add clients* to pyproject.toml tool.setuptools.packages.find include so 'pip install .' and the Docker layer pick up the new top-level package alongside the existing ones.

Pure move + import rewrite — no behavior change. Unit suite remains at 1987 passed.

Refs #394.
