+++ clients/bandcamp.py
@@ -14,7 +14,7 @@ import re
 
 import httpx
 
-from scripts.streaming_availability.http_client import BaseStreamingClient
+from clients.streaming.base import BaseStreamingClient
 
 log = logging.getLogger(__name__)
