+++ clients/streaming/base.py
