+++ clients/streaming/spotify.py
@@ -7,7 +7,7 @@ import base64
 import logging
 import time
 
-from scripts.streaming_availability.http_client import BaseStreamingClient
+from clients.streaming.base import BaseStreamingClient
 
 logger = logging.getLogger(__name__)
