+++ pyproject.toml
@@ -46,7 +46,7 @@ requires = ["setuptools>=65.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
-include = ["core*", "config*", "library*", "discogs*", "lookup*", "routers*", "services*", "generated*", "scripts*", "entity*"]
+include = ["core*", "config*", "library*", "discogs*", "lookup*", "routers*", "services*", "generated*", "scripts*", "entity*", "clients*"]
 
 [tool.ruff]
 line-length = 100
