+++ scripts/streaming_availability/dedup.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass, field
 import aiosqlite
 from wxyc_etl.text import is_compilation_artist
 
-from scripts.streaming_availability.matching import normalize_album_title, normalize_artist_name
+from clients.streaming.matching import normalize_album_title, normalize_artist_name
 
 
 @dataclass
