+++ streaming/dependencies.py
@@ -6,11 +6,11 @@ import logging
 
 from fastapi import Depends
 
+from clients.bandcamp import BandcampClient
+from clients.streaming.apple_music import AppleMusicClient
+from clients.streaming.deezer import DeezerClient
+from clients.streaming.spotify import SpotifyClient
 from config.settings import Settings, get_settings
-from scripts.bandcamp_client import BandcampClient
-from scripts.streaming_availability.apple_music_client import AppleMusicClient
-from scripts.streaming_availability.deezer_client import DeezerClient
-from scripts.streaming_availability.spotify_client import SpotifyClient
 
 logger = logging.getLogger(__name__)
